backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# Database
sqlalchemy[asyncio]==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6

# Gemini
google-generativeai==0.8.4

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.2

# LangChain (Phase 3 — agent)
langchain==0.3.14
langchain-google-genai==2.0.8

# Eval (Phase 5 — RAGAS)
ragas==0.2.8
datasets==3.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Utils
python-dotenv==1.0.1
pydantic==2.10.4
